aqualand/aqualand_app/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-chartjs==2.3.0
Pillow==12.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.10.0
cffi==2.0.0
pycparser==2.23
sqlparse==0.5.3
tzdata==2025.2
